requirements.txt
# Data & ETL
pandas>=2.0.0
numpy>=1.24.0
simfin>=0.9.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# App
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Notebooks (dev only)
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
